LayoutTests/imported/w3c/web-platform-tests/web-animations/timing-model/animations/canceling-an-animation-expected.txt


FAIL A play-pending ready promise should be rejected when the animation is canceled assert_equals: ready promise is rejected with AbortError expected "AbortError" but got "Error"
FAIL A pause-pending ready promise should be rejected when the animation is canceled assert_equals: ready promise is rejected with AbortError expected "AbortError" but got "Error"
PASS When an animation is canceled, it should create a resolved Promise 
FAIL The ready promise should be replaced when the animation is canceled assert_not_equals: got disallowed value object "[object Promise]"
PASS The finished promise should NOT be rejected if the animation is already idle 
PASS The cancel event should NOT be fired if the animation is already idle 
